requirements.txt
Django>=4.2.9,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.9,<3.0
drf-spectacular>=0.27.0,<0.28
Pillow>=10.2.0,<10.3.0
drf-extensions>=0.7.1,<0.8
